settimo commit

+++ S3/Project2-B-V/esercizio/src/app/movies/movie-list/movie-list.component.ts
@@ -1,16 +1,18 @@
 import { Component, OnInit } from '@angular/core';
 import { MovieService } from '../movie.service';
+import { Router } from '@angular/router';
 import { iMovie } from '../../models/i-movie';
+import { AuthService } from '../../auth/auth.service';
 
 @Component({
   selector: 'app-movie-list',
-  templateUrl: './movie-list.component.html',
+  templateUrl: './/movie-list.component.html',
   styleUrls: ['./movie-list.component.scss']
 })
 export class MovieListComponent implements OnInit {
   movies: iMovie[] = [];
 
-  constructor(private movieService: MovieService) {}
+  constructor(private movieService: MovieService, private router: Router,  private authService: AuthService) {}
 
   ngOnInit(): void {
     this.loadMovies();
@@ -29,16 +31,20 @@ export class MovieListComponent implements OnInit {
   }
 
   toggleFavorite(movie: iMovie): void {
-    if (movie.isFavourite) {
-      this.movieService.removeFavorite(movie.id).subscribe(() => {
-        movie.isFavourite = false;
-        console.log(`${movie.title} removed from favorites`);
-      });
+    if (this.authService.isAuthenticated()) {
+      if (movie.isFavourite) {
+        this.movieService.removeFavorite(movie.id).subscribe(() => {
+          movie.isFavourite = false;
+          console.log(`${movie.title} removed from favorites`);
+        });
+      } else {
+        this.movieService.addFavorite(movie).subscribe(() => {
+          movie.isFavourite = true;
+          console.log(`${movie.title} added to favorites`);
+        });
+      }
     } else {
-      this.movieService.addFavorite(movie).subscribe(() => {
-        movie.isFavourite = true;
-        console.log(`${movie.title} added to favorites`);
-      });
+      this.router.navigate(['/auth/login']);
     }
   }
 }
